Function for getting weather data from one station
Source: R/measures.R

Function for getting weather data from one station

riem_measures(station = "VOHY", date_start = "2014-01-01",
  date_end = as.character(Sys.Date()))

Arguments

station
station ID, see riem_stations()

date_start
date of start of the desired data, e.g. "2000-01-01"

date_end
date of end of the desired data, e.g. "2016-04-22"

Value

a data.frame (tibble tibble) with measures, the number of columns can vary from station to station, but possible variables are

station: three or four character site identifier

valid: timestamp of the observation (UTC)

tmpf: Air Temperature in Fahrenheit, typically @ 2 meters

dwpf: Dew Point Temperature in Fahrenheit, typically @ 2 meters

relh: Relative Humidity in

drct: Wind Direction in degrees from north

sknt: Wind Speed in knots

p01i: One hour precipitation for the period from the observation time to the time of the previous hourly precipitation reset. This varies slightly by site. Values are in inches. This value may or may not contain frozen precipitation melted by some device on the sensor or estimated by some other means. Unfortunately, we do not know of an authoritative database denoting which station has which sensor.

alti: Pressure altimeter in inches

mslp: Sea Level Pressure in millibar

vsby: Visibility in miles

gust: Wind Gust in knots

skyc1: Sky Level 1 Coverage

skyc2: Sky Level 2 Coverage

skyc3: Sky Level 3 Coverage

skyc4: Sky Level 4 Coverage

skyl1: Sky Level 1 Altitude in feet

skyl2: Sky Level 2 Altitude in feet

skyl3: Sky Level 3 Altitude in feet

skyl4: Sky Level 4 Altitude in feet

presentwx: Present Weather Codes (space seperated), see e.g. [this manual](http://www.ofcm.gov/fmh-1/pdf/H-CH8.pdf) for further explanations.

metar: unprocessed reported observation in METAR format

Details

The data is queried through https://mesonet.agron.iastate.edu/request/download.phtml.

Examples

# NOT RUN {
riem_measures(station = "VOHY", date_start = "2000-01-01", date_end = "2016-04-22")
# }