Function for getting weather data from one station

Function for getting weather data from one station

riem_measures(station = "VOHY", date_start = "2014-01-01",
  date_end = as.character(Sys.Date()))

Arguments

station
station ID, see riem_stations()

date_start
date of start of the desired data, e.g. "2000-01-01"

date_end
date of end of the desired data, e.g. "2016-04-22"

Value

a data.frame (tibble tibble) with measures, the number of columns can vary from station to station, but possible variables are

station: three or four character site identifier

valid: timestamp of the observation (UTC)

tmpf: Air Temperature in Fahrenheit, typically @ 2 meters

dwpf: Dew Point Temperature in Fahrenheit, typically @ 2 meters

relh: Relative Humidity in

drct: Wind Direction in degrees from north

sknt: Wind Speed in knots

p01i: One hour precipitation for the period from the observation time to the time of the previous hourly precipitation reset. This varies slightly by site. Values are in inches. This value may or may not contain frozen precipitation melted by some device on the sensor or estimated by some other means. Unfortunately, we do not know of an authoritative database denoting which station has which sensor.

alti: Pressure altimeter in inches

mslp: Sea Level Pressure in millibar

vsby: Visibility in miles

gust: Wind Gust in knots

skyc1: Sky Level 1 Coverage

skyc2: Sky Level 2 Coverage

skyc3: Sky Level 3 Coverage

skyc4: Sky Level 4 Coverage

skyl1: Sky Level 1 Altitude in feet

skyl2: Sky Level 2 Altitude in feet

skyl3: Sky Level 3 Altitude in feet

skyl4: Sky Level 4 Altitude in feet

presentwx: Present Weather Codes (space seperated), see e.g. Chapter 8 of [this manual](https://www.ofcm.gov/publications/fmh/FMH1/FMH1.pdf) for further explanations.

feel: Apparent Temperature (Wind Chill or Heat Index) in degF

ice_accretion_1hr: Ice Accretion over 1 Hour in inch

ice_accretion_3hr: Ice Accretion over 3 Hour in inch

ice_accretion_6hr: Ice Accretion over 6 Hour in inch

relh: Relative Humidity in

metar: unprocessed reported observation in METAR format

peak_wind_gust: Wind gust in knots from the METAR PK WND remark, this value may be different than the value found in the gust field. The gust field is derived from the standard METAR wind report.

peak_wind_drct: The wind direction in degrees North denoted in the METAR PK WND remark.

peak_wind_time: The timestamp of the PK WND value in the same timezone as the valid field and controlled by the tz parameter.

Details

The data is queried through https://mesonet.agron.iastate.edu/request/download.phtml.

Examples

# NOT RUN {
riem_measures(station = "VOHY", date_start = "2000-01-01", date_end = "2016-04-22")
# }